renamed userapi.ts to profilegetapi.ts and added profilepu tapi.ts
